# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Logistics Optimization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: getShipments
      tags: [shipments]
      summary: List all shipments
      responses:
        "200":
          description: List of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shipment"
    post:
      operationId: createShipment
      tags: [shipments]
      summary: Create a shipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentBody"
      responses:
        "201":
          description: Created shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
  /shipments/{id}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
    put:
      operationId: updateShipment
      tags: [shipments]
      summary: Update shipment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentBody"
      responses:
        "200":
          description: Updated shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"

  # ─── Routes ───────────────────────────────────────────────────────────────
...
  /routes/optimize:
    post:
      operationId: optimizeRoute
      tags: [routes]
      summary: Request mock AI route optimization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizeRouteBody"
      responses:
        "200":
          description: Optimized route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"

  # ─── Inventory ────────────────────────────────────────────────────────────
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List inventory items
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # ─── Fleet & Drivers ──────────────────────────────────────────────────────
...
  /fleet/assign:
    post:
      operationId: assignDriver
      tags: [fleet]
      summary: Assign driver to truck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDriverBody"
      responses:
        "200":
          description: Assignment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignDriverResult"

  # ─── Analytics ────────────────────────────────────────────────────────────
...
  /verify/scan:
    post:
      operationId: scanShipment
      tags: [verification]
      summary: Scan and verify shipment barcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanBody"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentVerification"
  /verify/photo:
    post:
      operationId: photoVerifyShipment
      tags: [verification]
      summary: Submit photo verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoVerifyBody"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentVerification"
  /verify/{shipmentId}:
    get:
      operationId: getVerification
      tags: [verification]
      summary: Get verification status for a shipment
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentVerification"
  /verify/report-damage:
    post:
      operationId: reportDamage
      tags: [verification]
      summary: Report shipment damage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DamageReportBody"
      responses:
        "200":
          description: Damage report result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentVerification"
...
components:
  schemas:
...
    Shipment:
      type: object
      properties:
        id:
          type: string
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
          enum: [pending, in_transit, delivered, delayed, cancelled]
        driver_id:
          type: string
          nullable: true
        estimated_delivery:
          type: string
        weight_kg:
          type: number
        priority:
          type: string
          enum: [low, medium, high]
        created_at:
          type: string
      required: [id, origin, destination, status, estimated_delivery, weight_kg, priority, created_at]
    ShipmentDetail:
      allOf:
        - $ref: "#/components/schemas/Shipment"
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/ShipmentEvent"
          required: [events]
    ShipmentEvent:
      type: object
      properties:
        id:
          type: string
        shipment_id:
          type: string
        event_type:
          type: string
        description:
          type: string
        timestamp:
          type: string
        location:
          type: string
      required: [id, shipment_id, event_type, description, timestamp]
    CreateShipmentBody:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        weight_kg:
          type: number
        priority:
          type: string
          enum: [low, medium, high]
        estimated_delivery:
          type: string
      required: [origin, destination, weight_kg, priority, estimated_delivery]
    UpdateShipmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_transit, delivered, delayed, cancelled]
        driver_id:
          type: string
      required: [status]

    # ── Routes ────────────────────────────────────────────────────────────
    Route:
      type: object
      properties:
        route_id:
          type: string
        shipment_id:
          type: string
          nullable: true
        origin:
          type: string
        destination:
          type: string
        distance_km:
          type: number
        estimated_time:
          type: string
        traffic_level:
          type: string
          enum: [low, moderate, high]
        optimized:
          type: boolean
        waypoints:
          type: array
          items:
            type: string
        fuel_savings_pct:
          type: number
        created_at:
          type: string
      required: [route_id, origin, destination, distance_km, estimated_time, traffic_level, optimized, waypoints, fuel_savings_pct, created_at]
    OptimizeRouteBody:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        shipment_id:
          type: string
      required: [origin, destination]

    # ── Inventory ─────────────────────────────────────────────────────────
    InventoryItem:
      type: object
      properties:
        id:
          type: string
        product_name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        warehouse:
          type: string
        category:
          type: string
        predicted_demand:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        confidence:
          type: string
        reorder_point:
          type: integer
        unit_price:
          type: number
        updated_at:
          type: string
      required: [id, product_name, sku, quantity, warehouse, category, predicted_demand, confidence, reorder_point, unit_price, updated_at]
    CreateInventoryBody:
      type: object
      properties:
        product_name:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        warehouse:
          type: string
        category:
          type: string
        reorder_point:
          type: integer
        unit_price:
          type: number
      required: [product_name, sku, quantity, warehouse, category, reorder_point, unit_price]
    UpdateInventoryBody:
      type: object
      properties:
        quantity:
          type: integer
        warehouse:
          type: string
      required: [quantity]

    # ── Fleet & Drivers ───────────────────────────────────────────────────
...
    AssignDriverBody:
      type: object
      properties:
        driver_id:
          type: string
        truck_id:
          type: string
      required: [driver_id, truck_id]
    AssignDriverResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        driver_id:
          type: string
        truck_id:
          type: string
      required: [success, message, driver_id, truck_id]

    # ── Analytics ─────────────────────────────────────────────────────────
...
    ShipmentVerification:
      type: object
      properties:
        id:
          type: string
        shipment_id:
          type: string
        scan_code:
          type: string
          nullable: true
        verification_status:
          type: string
          enum: [pending, passed, failed, damaged]
        image_url:
          type: string
          nullable: true
        damage_flag:
          type: boolean
        damage_description:
          type: string
          nullable: true
        verified_by:
          type: string
          nullable: true
        confidence:
          type: string
        timestamp:
          type: string
      required: [id, shipment_id, verification_status, damage_flag, confidence, timestamp]
    ScanBody:
      type: object
      properties:
        shipment_id:
          type: string
        scan_code:
          type: string
        verified_by:
          type: string
      required: [shipment_id, scan_code]
    PhotoVerifyBody:
      type: object
      properties:
        shipment_id:
          type: string
        image_url:
          type: string
        verified_by:
          type: string
      required: [shipment_id, image_url]
    DamageReportBody:
      type: object
      properties:
        shipment_id:
          type: string
        damage_description:
          type: string
        verified_by:
          type: string
      required: [shipment_id, damage_description]

    # ── Workforce ─────────────────────────────────────────────────────────
...