# lib/api-zod/src/generated/types/assignDriverBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logistics Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignDriverBody {
  driver_id: string;
  truck_id: string;
}