# lib/api-zod/src/generated/types/optimizeRouteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logistics Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OptimizeRouteBody {
  origin: string;
  destination: string;
  shipment_id?: string;
}