# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logistics Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteTrafficLevel } from "./routeTrafficLevel";

export interface Route {
  route_id: string;
  shipment_id?: string | null;
  origin: string;
  destination: string;
  distance_km: number;
  estimated_time: string;
  traffic_level: RouteTrafficLevel;
  optimized: boolean;
  waypoints: string[];
  fuel_savings_pct: number;
  created_at: string;
}