# lib/api-zod/src/generated/types/shipmentDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Logistics Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Shipment } from "./shipment";
import type { ShipmentEvent } from "./shipmentEvent";

export type ShipmentDetail = Shipment & {
  events: ShipmentEvent[];
};